Added Wikipedia functionality to map markers
As suggested I improved the app by adding the Wikipedia link call to the places markers. I took the AJAX call out of the setPlace function to make it available for the event listener that defines the marker click. I also, added the infowindow closing when filtering down the places to avoid lingering ghosts infowindows.